Optimize a function with the PB-Hybrid algorithm

This function implements a two-step, population-based, hybrid optimization algorithm, for solving problems of the form:

_x, x f(x, x) ~~s.t.~~ g(x, x) 0,~ g(x) 0\ x X,~ x X

See the readme for more details on this format of problems, and when a pb-hy algorithm would be useful.

Usage

optimize_phy(
  f = function(x, xtil) NA,
  g = function(x, xtil) 0,
  gtil = function(xtil) 0,
  x_dom,
  xtil_dom,
  initializer,
  optimizer,
  updater,
  stopper = flow_stopper(),
  logger = flow_logger(),
  check_samples = NULL,
  check_op = c(1, 2)
)

Arguments

f

Objective function f(x, x).

g, gtil

Constraint functions g(x, x) and g(x). By default, never-binding ones.

x_dom, xtil_dom

Domains X and X. These will be used by `optimizer` (for X), and `initializer`/`updater` (for X). Should be a list with m and m entries.

initializer

Operator (function) to initialize the population, such that S_0 = initializer(X, g) (see details).

optimizer

Operator (function) to solve the reduced problem, such that R_t = optimizer(X, g, t, S_t) (see details).

updater

Operator (function) to update the population, such that S_t+1 = updater(X, g, t, S_t, R_t).

stopper

Stopping criteria for the algorithm, created via [flow_stopper()].

logger

Logger for tracking the process, created via [flow_logger()].

check_samples

Number of samples to be expected in S_t. Set to `NULL` to not check (required if it changes with iterations t).

check_op

Integer vector with iterations to check if the operators are producing results with the needed format. Helps catching errors. Set to `0` to not check.

Value

A list containing:

results

Optimization results for each iteration.

metrics

Metrics calculated during the optimization process.

duration

Timing information for initialization, main loop, and total execution.

Details

This function is a wrapper for the operator functions, organizing them in the correct structure of a population-based, two-step, hybrid algorithm. But, it is the job of the user to implement the operators correctly: * `initializer`: todo. * `optimizer`: todo. * `updater`: todo.